Minor update regarding session file

+++ src/utils/session_handler.py
@@ -59,18 +59,22 @@ def flush(url):
   info_msg = "Flushing the stored session from the session file... "
   sys.stdout.write(settings.print_info_msg(info_msg))
   sys.stdout.flush()
-  try:
-    conn = sqlite3.connect(settings.SESSION_FILE)
-    tables = list(conn.execute("SELECT name FROM sqlite_master WHERE type is 'table'"))
-    conn.executescript(';'.join(["DROP TABLE IF EXISTS %s" %i for i in tables]))
-    conn.commit()
-    conn.close()
-    print "[ " + Fore.GREEN + "SUCCEED" + Style.RESET_ALL + " ]"
-
-  except sqlite3.OperationalError, err_msg:
+  if not os.path.isfile(settings.SESSION_FILE):
     print "[ " + Fore.RED + "FAILED" + Style.RESET_ALL + " ]"
-    err_msg = "Unable to flush the session file." + str(err_msg).title()
+    err_msg = "The session file does not exist."
     print settings.print_critical_msg(err_msg)
+  else:
+    try:
+      conn = sqlite3.connect(settings.SESSION_FILE)
+      tables = list(conn.execute("SELECT name FROM sqlite_master WHERE type is 'table'"))
+      conn.executescript(';'.join(["DROP TABLE IF EXISTS %s" %i for i in tables]))
+      conn.commit()
+      conn.close()
+      print "[ " + Fore.GREEN + "SUCCEED" + Style.RESET_ALL + " ]"
+    except sqlite3.OperationalError, err_msg:
+      print "[ " + Fore.RED + "FAILED" + Style.RESET_ALL + " ]"
+      err_msg = "Unable to flush the session file." + str(err_msg).title()
+      print settings.print_critical_msg(err_msg)    
 
 """
 Clear injection point records
